+++ lib/src/mixins/error_notifier.dart
@@ -1,3 +1,4 @@
+import 'package:bon_notifiers/src/notifiers/async_notifier.dart';
 import 'package:flutter/foundation.dart';
 
 /// Mixin for handling error state.
@@ -11,8 +12,11 @@ mixin ErrorNotifier on ChangeNotifier {
   bool get hasError => _error != null;
 
   /// Sets the error state and notifies listeners.
-  void setError(Object error) {
+  void setError(String message, Object error, [StackTrace? stackTrace]) {
     _error = error;
+    if (AsyncNotifier.errorListener != null) {
+      AsyncNotifier.errorListener!(message, error, this, stackTrace);
+    }
     notifyListeners();
   }