+++ lib/src/notifiers/async_notifier.dart
@@ -39,8 +39,11 @@ class AsyncNotifier<T> extends ChangeNotifier implements AsyncListenable<T> {
   bool get isLoading => _loading;
 
   /// Sets an [error] and notifies listeners.
-  void setError(Object error, [StackTrace? st]) {
+  void setError(String message, Object error, [StackTrace? st]) {
     _error = error;
+    if (errorListener != null) {
+      errorListener!(message, error, this, st);
+    }
     notifyListeners();
   }
 
@@ -60,4 +63,12 @@ class AsyncNotifier<T> extends ChangeNotifier implements AsyncListenable<T> {
     _loading = false;
     notifyListeners();
   }
+
+  static void Function(
+    String message,
+    Object error,
+    ChangeNotifier notifier,
+    StackTrace? stackTrace,
+  )?
+  errorListener;
 }